# server/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FoodInsight API",
    "description": "\n## Smart Snack Inventory Monitoring API\n\nLocal-first API for managing snack inventory on FoodInsight edge devices.\n\n### Architecture\n\n- **Local SQLite Database**: All data stored locally, no cloud dependencies\n- **Role-Based Access**: Public inventory endpoints, authenticated admin endpoints\n- **Real-Time Detection**: Receives inventory updates from edge detection pipeline\n\n### Authentication\n\nAdmin endpoints require HTTP Basic authentication:\n- `viewer`: Read-only access to status and configuration\n- `operator`: Can start/stop detection pipeline\n- `admin`: Full access including user management\n\n### Default Credentials\n\n| Username | Password | Role |\n|----------|----------|------|\n| admin | admin | admin |\n\n**Important**: Change default credentials after first login.\n        ",
    "version": "1.0.0",
    "x-logo": {
      "url": "https://example.com/logo.png"
    },
    "contact": {
      "name": "FoodInsight Support",
      "email": "support@foodinsight.local"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "paths": {
...
    "/inventory/update": {
      "post": {
        "tags": [
          "inventory"
        ],
        "summary": "Update Inventory",
        "description": "Update inventory from detection pipeline.\n\nCalled internally by the detection service after each frame analysis.\nNo authentication required for local device communication.",
        "operationId": "update_inventory_inventory_update_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InventoryUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/inventory/event": {
      "post": {
        "tags": [
          "inventory"
        ],
        "summary": "Log Inventory Event",
        "description": "Log a detection event.\n\nCalled by detection pipeline when inventory changes are detected.",
        "operationId": "log_inventory_event_inventory_event_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InventoryEventCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/inventory/item/{item_name}": {
      "get": {
        "tags": [
          "inventory"
        ],
        "summary": "Get Inventory Item",
        "description": "Get a single inventory item by name.",
        "operationId": "get_inventory_item_inventory_item__item_name__get",
        "parameters": [
          {
            "name": "item_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Item Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/inventory/events": {
      "get": {
        "tags": [
          "inventory"
        ],
        "summary": "Get Inventory Events",
        "description": "Get recent inventory events.\n\nPublic endpoint - no authentication required.",
        "operationId": "get_inventory_events_inventory_events_get",
        "parameters": [
          {
            "name": "event_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by event type",
              "title": "Event Type"
            },
            "description": "Filter by event type"
          },
          {
            "name": "item_name",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by item name",
              "title": "Item Name"
            },
            "description": "Filter by item name"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "InventoryEventCreate": {
        "properties": {
          "event_type": {
            "type": "string",
            "title": "Event Type"
          },
          "item_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Item Name"
          },
          "count_before": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Count Before"
          },
          "count_after": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Count After"
          },
          "confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence"
          },
          "details": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          }
        },
        "type": "object",
        "required": [
          "event_type"
        ],
        "title": "InventoryEventCreate",
        "description": "Event to log with inventory update."
      },
      "InventoryItemUpdate": {
        "properties": {
          "count": {
            "type": "integer",
            "title": "Count"
          },
          "confidence": {
            "type": "number",
            "title": "Confidence",
            "default": 1.0
          }
        },
        "type": "object",
        "required": [
          "count"
        ],
        "title": "InventoryItemUpdate",
        "description": "Single item update from detection pipeline."
      },
      "InventoryUpdateRequest": {
        "properties": {
          "items": {
            "additionalProperties": {
              "$ref": "#/components/schemas/InventoryItemUpdate"
            },
            "type": "object",
            "title": "Items"
          },
          "timestamp": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timestamp"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "InventoryUpdateRequest",
        "description": "Payload for inventory update from detection pipeline."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBasic": {
        "type": "http",
        "scheme": "basic"
      }
    }
  },
...
}